+++ readux/books/urls.py
@@ -37,4 +37,7 @@ urlpatterns = patterns('',
 
     url(r'^(?P<vol_pid>[^/]+)/pages/(?P<pid>[^/]+)/(?P<mode>(thumbnail|mini-thumbnail|single-page|fs|info))/$',
         views.PageImage.as_view(),  name='page-image'),
+    url(r'^(?P<vol_pid>[^/]+)/pages/(?P<pid>[^/]+)/(?P<mode>iiif)(?P<url>.*)$',
+        views.PageImage.as_view())
+
 )