+++ readux/books/views.py
@@ -11,8 +11,10 @@ from django.views.decorators.http import condition, require_http_methods, \
 from django.views.decorators.vary import vary_on_cookie
 from django.views.generic import ListView, DetailView, View
 from django.views.generic.base import RedirectView
+import json
 from urllib import urlencode
 import os
+import requests
 import logging
 
 from eulfedora.server import Repository, TypeInferringRepository, RequestFailed
@@ -22,7 +24,7 @@ from readux.books.models import Volume, SolrVolume, Page, VolumeV1_0, \
     PageV1_1, SolrPage
 from readux.books.forms import BookSearch
 from readux.books import view_helpers, annotate
-from readux.utils import solr_interface
+from readux.utils import solr_interface, absolutize_url
 from readux.views import VaryOnCookieMixin
 
 
@@ -603,7 +605,60 @@ class PageRedirect(RedirectView):
         return ''.join([page_url, kwargs['path']])
 
 
-class PageImage(RedirectView):
+class ProxyView(View):
+    # quick and dirty proxyview modeled on RedirectView
+
+    def get(self, request, *args, **kwargs):
+        url = self.get_redirect_url(*args, **kwargs)
+        # use headers to allow browsers to cache downloaded copies
+        headers = {}
+        for header in ['HTTP_IF_MODIFIED_SINCE', 'HTTP_IF_UNMODIFIED_SINCE',
+                       'HTTP_IF_MATCH', 'HTTP_IF_NONE_MATCH']:
+            if header in request.META:
+                headers[header.replace('HTTP_', '')] = request.META.get(header)
+        remote_response = requests.get(url, headers=headers)
+        local_response = HttpResponse()
+        local_response.status_code = remote_response.status_code
+
+        # include response headers, except for server-specific items
+        for header, value in remote_response.headers.iteritems():
+            if header not in ['Connection', 'Server', 'Keep-Alive', 'Link']:
+                             # 'Access-Control-Allow-Origin', 'Link']:
+                # FIXME: link header is valuable, but would
+                # need to be made relative to current url
+                local_response[header] = value
+
+        # special case, for deep zoom (hack)
+        if kwargs['mode'] == 'info':
+            data = remote_response.json()
+            # need to adjust the id to be relative to current url
+            # this is a hack, patching in a proxy iiif interface at this url
+            data['@id'] = absolutize_url(request.path.replace('/info/', '/iiif'))
+            local_response.content = json.dumps(data)
+            # upate content-length for change in data
+            local_response['content-length'] = len(local_response.content)
+        else:
+            # include response content if any
+            local_response.content = remote_response.content
+
+        return local_response
+
+    def head(self, request, *args, **kwargs):
+        url = self.get_redirect_url(*args, **kwargs)
+        remote_response = requests.head(url)
+        response = HttpResponse()
+        for header, value in remote_response.headers.iteritems():
+            if header not in ['Connection', 'Server', 'Keep-Alive',
+                             'Access-Control-Allow-Origin', 'Link']:
+                response[header] = value
+        return response
+
+
+# class PageImage(RedirectView):
+# NOTE: previously, was redirecting to loris, but currently the loris
+# image server is not externally accessible
+
+class PageImage(ProxyView):
     '''Local view for page images.  These all return redirects to the
     configured IIIF image viewer, but allow for a local, semantic
     image url independent of image handling implementations
@@ -612,6 +667,7 @@ class PageImage(RedirectView):
     def get_redirect_url(self, *args, **kwargs):
         repo = TypeInferringRepository()
         page = repo.get_object(kwargs['pid'], type=Page)
+
         if kwargs['mode'] == 'thumbnail':
             return page.iiif.thumbnail()
         elif kwargs['mode'] == 'mini-thumbnail':
@@ -621,4 +677,8 @@ class PageImage(RedirectView):
         elif kwargs['mode'] == 'fs':  # full size
             return page.iiif
         elif kwargs['mode'] == 'info':
+            # TODO: needs an 'Access-Control-Allow-Origin' header
+            # to allow jekyll sites to use for deep zoom
             return page.iiif.info()
+        elif kwargs['mode'] == 'iiif':
+            return page.iiif.info().replace('info.json', kwargs['url'].rstrip('/'))